+++ source/include/G3D/Cylinder.h
@@ -34,7 +34,7 @@ public:
     /** Uninitialized */
     Cylinder();
     Cylinder(class BinaryInput& b);
-	Cylinder(const Vector3& _p1, const Vector3& _p2, double _r);
+	Cylinder(const Vector3& _p1, const Vector3& _p2, float _r);
 	void serialize(class BinaryOutput& b) const;
 	void deserialize(class BinaryInput& b);
